[Epic]: Custom Connector Implementations in the Connector Service #2128 - merged master refactoring in
