+++ lib/utility/sink.rb
